refactor(nextly): move the single companion reconcile out of the handler
The companion provisioning step was a private function inside the request
handler, so every other path that changes a Single's schema would have needed
its own copy. It moves beside the Singles services unchanged: same arguments,
same statements, same best-effort runtime registration.
